requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
asn1crypto=0.24.0=py37_1003
blinker=1.4=py_1
bzip2=1.0.6=h1de35cc_1002
ca-certificates=2019.3.9=hecc5488_0
cachetools=2.1.0=py_0
certifi=2019.3.9=py37_0
cffi=1.12.2=py37h2d6ddff_1
chardet=3.0.4=py37_1003
cryptography=2.5=py37hc2b1221_1
google-api-python-client=1.7.8=py_0
google-auth=1.6.3=py_0
google-auth-httplib2=0.0.3=py_2
google-auth-oauthlib=0.2.0=py37_1000
httplib2=0.12.1=py37_0
idna=2.8=py37_1000
libcxx=7.0.0=h2d50403_1
libffi=3.2.1=h0a44026_1005
llvm-meta=7.0.0=0
ncurses=6.1=h0a44026_1002
oauthlib=3.0.1=py_0
openssl=1.1.1b=h1de35cc_1
pip=19.0.3=py37_0
pyasn1=0.4.4=py_1
pyasn1-modules=0.2.4=py_0
pycparser=2.19=py_0
pyjwt=1.7.1=py_0
pyopenssl=19.0.0=py37_0
pysocks=1.6.8=py37_1002
python=3.7.1=hbdd33cc_1002
readline=7.0=hcfe32e1_1001
requests=2.21.0=py37_1000
requests-oauthlib=1.2.0=py_0
rsa=3.4.2=py_1
setuptools=40.8.0=py37_0
simplejson=3.16.0=py37h1de35cc_1002
six=1.12.0=py37_1000
sqlite=3.26.0=h1765d9f_1001
tk=8.6.9=ha441bb4_1000
uritemplate=3.0.0=py_1
urllib3=1.24.1=py37_1000
wheel=0.33.1=py37_0
xz=5.2.4=h1de35cc_1001
zlib=1.2.11=h1de35cc_1004
